+++ app/src/main/java/org/wit/archaeologicalfieldwork/activities/hillfort/HillfortListActivity.kt
@@ -1,7 +1,6 @@
 package org.wit.archaeologicalfieldwork.activities.hillfort
 
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
 import android.view.*
@@ -11,10 +10,12 @@ import org.jetbrains.anko.intentFor
 import org.jetbrains.anko.startActivityForResult
 import org.wit.archaeologicalfieldwork.R
 import org.wit.archaeologicalfieldwork.activities.HomeActivity
-import org.wit.archaeologicalfieldwork.activities.profile.UserActivity
-import org.wit.archaeologicalfieldwork.activities.profile.userLogged
+import org.wit.archaeologicalfieldwork.adapters.HillfortAdapter
+import org.wit.archaeologicalfieldwork.adapters.HillfortListener
+import org.wit.archaeologicalfieldwork.activities.user.UserActivity
+import org.wit.archaeologicalfieldwork.activities.user.userLogged
 import org.wit.archaeologicalfieldwork.main.MainApp
-import org.wit.archaeologicalfieldwork.models.HillfortModel
+import org.wit.archaeologicalfieldwork.models.hillfort.HillfortModel
 
 
 class HillfortListActivity : HomeActivity(), HillfortListener {
@@ -32,30 +33,14 @@ class HillfortListActivity : HomeActivity(), HillfortListener {
 
     val contentView = layoutInflater.inflate(R.layout.activity_hillfort_list, null, false)
     drawer_layout.addView(contentView, 0)
-    toolbarMain.title = title
-    setSupportActionBar(toolbarMain)
 
     val layoutManager = LinearLayoutManager(this)
     recyclerView.layoutManager = layoutManager
     loadHillforts()
   }
 
-  override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-    menuInflater.inflate(R.menu.menu_main, menu)
-    return super.onCreateOptionsMenu(menu)
-  }
-
-  override fun onOptionsItemSelected(item: MenuItem?): Boolean {
-    when (item?.itemId) {
-      R.id.item_add -> startActivityForResult<HillfortActivity>(0)
-      R.id.profile -> startActivityForResult<UserActivity>(0)
-    }
-    return super.onOptionsItemSelected(item)
-  }
-
   override fun onHillfortClick(hillfort: HillfortModel) {
     startActivityForResult(intentFor<HillFortProfileActivity>().putExtra("hillfort", hillfort), 0)
-    //startActivityForResult<HillFortProfileActivity>(0)
   }
 
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
