+++ app/src/main/java/org/wit/archaeologicalfieldwork/main/MainApp.kt
@@ -3,10 +3,10 @@ package org.wit.archaeologicalfieldwork.main
 import android.app.Application
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
-import org.wit.archaeologicalfieldwork.models.HillfortJSONStore
-import org.wit.archaeologicalfieldwork.models.HillfortStore
-import org.wit.archaeologicalfieldwork.models.UserJSONStore
-import org.wit.archaeologicalfieldwork.models.UserStore
+import org.wit.archaeologicalfieldwork.models.hillfort.HillfortJSONStore
+import org.wit.archaeologicalfieldwork.models.hillfort.HillfortStore
+import org.wit.archaeologicalfieldwork.models.user.UserJSONStore
+import org.wit.archaeologicalfieldwork.models.user.UserStore
 
 class MainApp : Application(), AnkoLogger {
