--- app/src/main/java/org/wit/archaeologicalfieldwork/models/UserModel.kt
@@ -1,9 +1,0 @@
-package org.wit.archaeologicalfieldwork.models
-
-import android.annotation.SuppressLint
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@SuppressLint("ParecelCreator")
-@Parcelize
-data class UserModel(var id: Long = 0,var name: String = "", var email: String = "", var password: String = "", var stats: MutableList<StatsModel> = mutableListOf(), var userImage: String = "", var joined: String = "") : Parcelable
