+++ app/src/main/java/org/wit/archaeologicalfieldwork/models/comment/CommentsModel.kt
@@ -1,4 +1,4 @@
-package org.wit.archaeologicalfieldwork.models
+package org.wit.archaeologicalfieldwork.models.comment
 
 import android.annotation.SuppressLint
 import android.os.Parcelable
