+++ app/src/main/java/org/wit/archaeologicalfieldwork/models/stats/StatsModel.kt
@@ -1,4 +1,4 @@
-package org.wit.archaeologicalfieldwork.models
+package org.wit.archaeologicalfieldwork.models.stats
 
 import android.annotation.SuppressLint
 import android.os.Parcelable
