+++ app/src/main/java/org/wit/archaeologicalfieldwork/models/user/UserModel.kt
@@ -1,0 +1,10 @@
+package org.wit.archaeologicalfieldwork.models.user
+
+import android.annotation.SuppressLint
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+import org.wit.archaeologicalfieldwork.models.stats.StatsModel
+
+@SuppressLint("ParecelCreator")
+@Parcelize
+data class UserModel(var id: Long = 0, var name: String = "", var email: String = "", var password: String = "", var stats: MutableList<StatsModel> = mutableListOf(), var userImage: String = "", var joined: String = "") : Parcelable
