+++ app/src/main/java/org/wit/archaeologicalfieldwork/models/user/UserStore.kt
@@ -1,4 +1,7 @@
-package org.wit.archaeologicalfieldwork.models
+package org.wit.archaeologicalfieldwork.models.user
+
+import org.wit.archaeologicalfieldwork.models.stats.StatsModel
+import org.wit.archaeologicalfieldwork.models.user.UserModel
 
 interface UserStore{
   fun create(user: UserModel)
