+++ app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".activities.profile.ProfileActivity">
+  tools:context=".activities.user.ProfileActivity">
 
   <RelativeLayout
     android:layout_width="match_parent"
