+++ app/src/main/res/layout/activity_user.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".activities.profile.UserActivity">
+  tools:context=".activities.user.UserActivity">
 
   <RelativeLayout
     android:layout_width="match_parent"
