+++ app/src/main/res/menu/drawer_view.xml
@@ -17,5 +17,9 @@
       android:id="@+id/nav_settings"
       android:icon="@drawable/baseline_settings_white_18dp"
       android:title="@string/menu_settings"/>
+    <item
+      android:id="@+id/nav_logout"
+      android:icon="@drawable/baseline_close_white_18dp"
+      android:title="@string/menu_logout"/>
   </group>
 </menu>
