--- app/src/main/res/menu/menu_main.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:tools="http://schemas.android.com/tools"
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto">
-
-  <item
-    android:id="@+id/profile"
-    android:icon="@android:drawable/ic_menu_view"
-    android:title="@string/menu_profile"
-    app:showAsAction="always" />
-
-  <item
-    android:id="@+id/item_add"
-    android:icon="@android:drawable/ic_menu_add"
-    android:title="@string/menu_addHillfort"
-    app:showAsAction="always" />
-</menu>
