--- app/src/main/res/menu/menu_profile.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:tools="http://schemas.android.com/tools"
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto">
-
-  <item
-    android:id="@+id/item_settings"
-    android:icon="@drawable/baseline_settings_white_18dp"
-    android:title="@string/menu_settings"
-    app:showAsAction="always" />
-
-  <item
-    android:id="@+id/item_logout"
-    android:icon="@drawable/baseline_settings_backup_restore_white_18dp"
-    android:title="@string/menu_logout"
-    app:showAsAction="always" />
-</menu>
